Bring the code from ebird-checklists into the project
Originally ebird-checklists was intended to be a base app which
provided the models and queries for other projects. However each
project is different, and keeping the projects in sync for every
change proved to be awkward, particularly since ebird-checklists was
in active development with changes from two projects that used it
being incorporated on a regular basis. The whole thing became
unmanageable, so ebird-checklists was changed from being a reuable
django app to a project starter, leaving each project that used it,
free to make the changes it needs.